+++ src/Newbe.BookmarkManager/Services/Services/Impl/TokenBaseInitializer.cs
@@ -1,0 +1,20 @@
+﻿using System.Net.Http.Headers;
+using Google.Apis.Http;
+
+namespace Newbe.BookmarkManager.Services
+{
+    public class TokenBaseInitializer : IConfigurableHttpClientInitializer
+    {
+        private readonly string _token;
+
+        public TokenBaseInitializer(string token)
+        {
+            _token = token;
+        }
+
+        public void Initialize(ConfigurableHttpClient httpClient)
+        {
+            httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", _token);
+        }
+    }
+}
